feat: 사전식물 삭제 구현

+++ backend/pium/src/main/java/com/official/pium/service/DictionaryPlantService.java
@@ -1,6 +1,8 @@
 package com.official.pium.service;
 
+import com.official.pium.domain.Admin;
 import com.official.pium.domain.DictionaryPlant;
+import com.official.pium.exception.NeedAdminException;
 import com.official.pium.mapper.DictionaryPlantMapper;
 import com.official.pium.repository.DictionaryPlantRepository;
 import com.official.pium.service.dto.DataResponse;
@@ -10,6 +12,7 @@ import com.official.pium.service.dto.DictionaryPlantSearchResponse;
 import com.official.pium.service.dto.DictionaryPlantUpdateRequest;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,17 +44,47 @@ public class DictionaryPlantService {
     }
 
     @Transactional
-    public void create(DictionaryPlantCreateRequest request) {
+    public void create(Admin admin, DictionaryPlantCreateRequest request) {
+        validateAdmin(admin);
         DictionaryPlant dictionaryPlant = DictionaryPlantMapper.toDictionaryPlant(request);
         dictionaryPlantRepository.save(dictionaryPlant);
     }
 
-    public void delete(Long id) {
-        // TODO: ADMIN
+    public void update(Admin admin, Long id, DictionaryPlantUpdateRequest request) {
+        validateAdmin(admin);
+        DictionaryPlant dictionaryPlant = dictionaryPlantRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("사전 식물이 존재하지 않습니다. id: "+ id));
+
+        dictionaryPlant.updateDictionaryPlant(
+                request.getName(),
+                request.getImageUrl(),
+                request.getFamilyName(),
+                request.getSmell(),
+                request.getPoison(),
+                request.getManageLevel(),
+                request.getGrowSpeed(),
+                request.getRequireTemp(),
+                request.getMinimumTemp(),
+                request.getRequireHumidity(),
+                request.getPostingPlace(),
+                request.getSpecialManageInfo(),
+                request.getSpring(),
+                request.getSummer(),
+                request.getAutumn(),
+                request.getWinter()
+        );
+    }
+
+    @Transactional
+    public void delete(Admin admin, Long id) {
+        validateAdmin(admin);
         dictionaryPlantRepository.deleteById(id);
     }
 
-    public void update(DictionaryPlantUpdateRequest request) {
 
+    private void validateAdmin(Admin admin) {
+        if (admin == null) {
+            throw new NeedAdminException("관리자 권한이 필요한 작업입니다.");
+        }
     }
 }
